requirements.txt
gradio==5.8.0
matplotlib==3.9.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.1.0
pydicom==3.0.1
pytorch-lightning==2.4.0
torchvision==0.21.0
